+++ operations/eu-central-1/env-kops/provider.tf
@@ -1,0 +1,6 @@
+# AWS region must be set externally, e.g. by exporting AWS_DEFAULT_REGION env variable
+
+provider "aws" {
+  region  = "${var.region}"
+  version = "~> 1.60"
+}