+++ CHANGELOG.md
@@ -5,7 +5,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
+### Changed
+- Adjustments to config repo change
+   ```
+   operations/$REGION/env -> operations/$REGION/env-kops
+                          -> operations/$REGION/env-eks
+   
+   application/$REGION/env -> application/$REGION/env-kops
+                           -> application/$REGION/env-eks
+   ```
 ## [0.3.0] - 2019-03-07
 ### Changed
 - Added input variable in operations for no_proxy configuration